generate_2fa/generate-2fa/requirements.txt
pyotp
qrcode[pil]
cryptography
psycopg2-binary
